Added 5 folders in a recurive way
