+++ DicomPseudonimyzer/__init__.py
@@ -1,0 +1,2 @@
+import DicomPseudonimyzer.anonymizer
+import DicomPseudonimyzer.simple_dicomanonymizer
