+++ docs/index.rst
@@ -4,7 +4,6 @@ Hello I'm a title
 Hello I'm an API
 
 .. toctree::
-   :maxdepth: 2
    :caption: Getting Started
 
    Introduction
